+++ docs/extend/config.md
@@ -5,7 +5,6 @@ aliases: [
 title: "Plugin config"
 description: "How develop and use a plugin with the managed plugin system"
 keywords: "API, Usage, plugins, documentation, developer"
-advisory: "experimental"
 ---
 
 <!-- This file is maintained within the docker/docker Github
@@ -19,8 +18,8 @@ advisory: "experimental"
 
 # Plugin Config Version 0 of Plugin V2
 
-This document outlines the format of the V0 plugin config. The plugin
-config described herein was introduced in the Docker daemon (experimental version) in the [v1.12.0
+This document outlines the format of the V0 plugin configuration. The plugin
+config described herein was introduced in the Docker daemon in the [v1.12.0
 release](https://github.com/docker/docker/commit/f37117045c5398fd3dca8016ea8ca0cb47e7312b).
 
 Plugin configs describe the various constituents of a docker plugin. Plugin
@@ -171,7 +170,6 @@ Config provides the base accessible fields for working with V0 plugin format
 
 ```
 {
-       	"configVersion": "v0",
        	"description": "A test plugin for Docker",
        	"documentation": "https://docs.docker.com/engine/extend/plugins/",
        	"entrypoint": ["plugin-no-remove", "/data"],
