+++ docs/extend/plugin_api.md
@@ -19,8 +19,7 @@ Docker plugins are out-of-process extensions which add capabilities to the
 Docker Engine.
 
 This document describes the Docker Engine plugin API. To view information on
-plugins managed by Docker Engine currently in experimental status, refer to
-[Docker Engine plugin system](index.md).
+plugins managed by Docker Engine, refer to [Docker Engine plugin system](index.md).
 
 This page is intended for people who want to develop their own Docker plugin.
 If you just want to learn about or use Docker plugins, look
